requirements.txt
django
djangorestframework
drfdocs
django-oauth-toolkit